+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
 - Initial version
+
+## 0.1.2
+
+- Send/Receive data
+- Send multiple packets
+- Send data with selectiveACK
+- Send keepalive
+- Anble to communicate with libutp